Add `--wheel` option to spin some names
I had a dream last night where this option existed, and then proceeded
to implement it when I woke up.

It made a lot of sense when I was writing it while I was slowly waking
up. Now a few hours later it makes no sense at all.

I'm keeping it anyway. You only live once. The new `--wheel` option
makes `qngng` spin an imaginary carnival wheel of random Queb names
honoring the other options. You see the names scrolling by as the wheel
spins and eventually it stops spinning and you're left with a winning
name.

The way I see it, it's a fun game to play during a family gathering.
Everyone sits in front of a screen where a terminal executing `qngng -w`
is projected, and the guests must guess the winning name. You can make
the game more interesting with categories and genders, for example

    $ qngng -wcsn

Have fun!

Also: I'm pretty sure this won't work so great on Windows..!

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>